# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CredentialUploadInput {
  /** Sanitized browser file name */
  file: string;
  fileSize: number;
  mimeType: string;
  /** @nullable */
  fileHash?: string | null;
}

export interface UploadReceipt {
  uploadId: string;
  fileName: string;
  fileSize: number;
  mimeType: string;
  receivedAt: string;
}

export interface AnalyzeCredentialInput {
  uploadId: string;
}

export type AnalysisJobStatus = typeof AnalysisJobStatus[keyof typeof AnalysisJobStatus];


export const AnalysisJobStatus = {
  queued: 'queued',
  processing: 'processing',
  complete: 'complete',
} as const;

export type AnalysisStepStatus = typeof AnalysisStepStatus[keyof typeof AnalysisStepStatus];


export const AnalysisStepStatus = {
  pending: 'pending',
  processing: 'processing',
  complete: 'complete',
} as const;

export interface AnalysisStep {
  key: string;
  label: string;
  status: AnalysisStepStatus;
}

export interface AnalysisJob {
  jobId: string;
  uploadId: string;
  status: AnalysisJobStatus;
  steps: AnalysisStep[];
}

export interface VerifyCredentialInput {
  certificateId: string;
  uploadId?: string;
  /** @nullable */
  fileHash?: string | null;
}

export type CredentialRecordStatus = typeof CredentialRecordStatus[keyof typeof CredentialRecordStatus];


export const CredentialRecordStatus = {
  verified: 'verified',
  suspicious: 'suspicious',
  invalid: 'invalid',
} as const;

export interface CredentialRecord {
  certificateId: string;
  studentName: string;
  institution: string;
  degree: string;
  registrationNumber: string;
  issueDate: string;
  graduationYear: number;
  status: CredentialRecordStatus;
}

export type VerificationSignalStatus = typeof VerificationSignalStatus[keyof typeof VerificationSignalStatus];


export const VerificationSignalStatus = {
  passed: 'passed',
  attention: 'attention',
  failed: 'failed',
  notChecked: 'notChecked',
} as const;

export interface VerificationSignal {
  key: string;
  label: string;
  status: VerificationSignalStatus;
  detail: string;
  points?: number;
  maxPoints?: number;
}

export type RiskFindingSeverity = typeof RiskFindingSeverity[keyof typeof RiskFindingSeverity];


export const RiskFindingSeverity = {
  LOW: 'LOW',
  MEDIUM: 'MEDIUM',
  HIGH: 'HIGH',
} as const;

export interface RiskFinding {
  type: string;
  severity: RiskFindingSeverity;
  message: string;
}

export type VerificationReportStatus = typeof VerificationReportStatus[keyof typeof VerificationReportStatus];


export const VerificationReportStatus = {
  verified: 'verified',
  likelyAuthentic: 'likelyAuthentic',
  suspicious: 'suspicious',
  invalid: 'invalid',
} as const;

export interface VerificationReport {
  certificate: CredentialRecord;
  score: number;
  status: VerificationReportStatus;
  timestamp: string;
  signals: VerificationSignal[];
  findings: RiskFinding[];
  documentHash: string;
  explanation: string[];
}

export type RecentCredentialStatus = typeof RecentCredentialStatus[keyof typeof RecentCredentialStatus];


export const RecentCredentialStatus = {
  verified: 'verified',
  suspicious: 'suspicious',
  invalid: 'invalid',
} as const;

export interface RecentCredential {
  certificateId: string;
  studentName: string;
  institution: string;
  score: number;
  status: RecentCredentialStatus;
  verifiedAt: string;
}

export type DashboardStatsWeeklyItem = {
  label: string;
  count: number;
};

export interface DashboardStats {
  total: number;
  verified: number;
  suspicious: number;
  invalid: number;
  weekly: DashboardStatsWeeklyItem[];
}