# lib/api-zod/src/generated/types/analysisStep.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisStepStatus } from './analysisStepStatus';

export interface AnalysisStep {
  key: string;
  label: string;
  status: AnalysisStepStatus;
}