# lib/api-zod/src/generated/types/credentialRecord.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CredentialRecordStatus } from './credentialRecordStatus';

export interface CredentialRecord {
  certificateId: string;
  studentName: string;
  institution: string;
  degree: string;
  registrationNumber: string;
  issueDate: string;
  graduationYear: number;
  status: CredentialRecordStatus;
}