# lib/api-zod/src/generated/types/credentialUploadInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CredentialUploadInput {
  /** Sanitized browser file name */
  file: string;
  fileSize: number;
  mimeType: string;
  /** @nullable */
  fileHash?: string | null;
}