# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardStatsWeeklyItem } from './dashboardStatsWeeklyItem';

export interface DashboardStats {
  total: number;
  verified: number;
  suspicious: number;
  invalid: number;
  weekly: DashboardStatsWeeklyItem[];
}