# lib/api-zod/src/generated/types/recentCredential.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RecentCredentialStatus } from './recentCredentialStatus';

export interface RecentCredential {
  certificateId: string;
  studentName: string;
  institution: string;
  score: number;
  status: RecentCredentialStatus;
  verifiedAt: Date;
}