# lib/api-zod/src/generated/types/riskFinding.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RiskFindingSeverity } from './riskFindingSeverity';

export interface RiskFinding {
  type: string;
  severity: RiskFindingSeverity;
  message: string;
}