# lib/api-zod/src/generated/types/verificationReport.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CredentialRecord } from './credentialRecord';
import type { RiskFinding } from './riskFinding';
import type { VerificationReportStatus } from './verificationReportStatus';
import type { VerificationSignal } from './verificationSignal';

export interface VerificationReport {
  certificate: CredentialRecord;
  score: number;
  status: VerificationReportStatus;
  timestamp: Date;
  signals: VerificationSignal[];
  findings: RiskFinding[];
  documentHash: string;
  explanation: string[];
}