# lib/api-zod/src/generated/types/verifyCredentialInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VerifyCredentialInput {
  certificateId: string;
  uploadId?: string;
  /** @nullable */
  fileHash?: string | null;
}